Change jira workdir to /opt/jira because of permissions.
Because /opt/jira-home is a volume it starts out owned by root and
with only root having access to it. So when the working directory
is set to /opt/jira-home and the user is not root we can't launch.
Setting the workdir to /opt/jira solves the issues since we use a
sudo script during launch to take ownership of /opt/jira-home and
so become the owner of /opt/jira-home.